optimised hw5a but this fails branch trace testcase
In this implementation when branch is taken the instruction in decode stage is flushed only when branch target offset isn't 4 (as then the next instruction in branch loaded will be the same as the one in decode making the flushing redundant. For the same reason instruction in fetch is only flushed if the offset isn't 4 as fetch already loaded next instruction to decode and also if the offset isn't 8 as this instruction is already loaded in fetch stage.